feat: add peer review changes
