+++ app/internal/agenda_events.py
@@ -58,59 +58,41 @@ def get_time_delta_string(start: date, end: date) -> str:
     return duration_string
 
 
-def _filter_dates(
+def filter_dates(
         events: List[Event],
         start: Union[None, date] = None,
         end: Union[None, date] = None,
 ) -> Iterator[Event]:
-    """filter events by a time frame."""
+    """Returns all events in a time frame.
+    if "start_date" or "end_date" are None,
+    it will ignore that parameter when filtering.
+
+    for example:
+    if start_date = None and end_date = datetime.now().date,
+    then the function will return all events that ends before end_date."""
 
     for event in events:
         if start and end:
             if start <= event.start.date() <= end:
                 yield event
+
         elif start:
             if start <= event.start.date():
                 yield event
+
         elif end:
             if end >= event.start.date():
                 yield event
+        else:
+            yield event
 
 
-def filter_dates(
-        events: List[Event], start: date,
-        end: date) -> Iterator[Event]:
-    """filter events by a time frame."""
-
-    yield from (
-        event for event in events
-        if (
-            start <= event.start.date() <= end
-            or start <= event.end.date() <= end
-            or (start >= event.start.date() and end <= event.end.date())
-        )
-    )
-
-
-def filter_start_dates(
-        events: Iterator[Event],
-        start: date,
+def get_events_in_time_frame(
+        start_date: Union[date, None],
+        end_date: Union[date, None],
+        user_id: int, db: Session
 ) -> Iterator[Event]:
-    """filter events by a time frame."""
+    """Yields all user's events in a time frame."""
 
-    yield from (
-        event for event in events
-        if start <= event.start.date()
-    )
-
-
-def filter_end_dates(
-        events: List[Event],
-        end: date
-) -> Iterator[Event]:
-    """filter events by a time frame."""
-
-    yield from (
-        event for event in events
-        if event.end.date() <= end
-    )
+    events = get_all_user_events(db, user_id)
+    yield from filter_dates(events, start_date, end_date)
