+++ tests/test_export.py
@@ -3,10 +3,12 @@ from datetime import timedelta
 from icalendar import vCalAddress
 
 from app.config import ICAL_VERSION, PRODUCT_ID
+from app.internal.agenda_events import filter_dates
 from app.routers.export import (
     create_ical_calendar, create_ical_event,
-    event_to_ical, get_events_in_time_frame
+    event_to_ical
 )
+from app.routers.user import get_all_user_events
 
 
 class TestExport:
@@ -21,18 +23,20 @@ class TestExport:
             next_month_event,
             sender, session
     ):
-        start = ''
-        end = ''
-        evens = get_events_in_time_frame(start, end, sender.id, session)
+        start = None
+        end = None
+        events = get_all_user_events(session, sender.id)
+        evens = filter_dates(events, start, end)
         assert set(evens) == {today_event, next_month_event}
 
     def test_get_events_end_time_frame_success(
             self, today_event,
             sender, session
     ):
-        start = ''
+        start = None
         end = today_event.end.date()
-        evens = get_events_in_time_frame(start, end, sender.id, session)
+        events = get_all_user_events(session, sender.id)
+        evens = filter_dates(events, start, end)
         assert list(evens) == [today_event]
 
     def test_get_events_end_time_frame_failure(
@@ -40,9 +44,10 @@ class TestExport:
             next_month_event,
             sender, session
     ):
-        start = ''
+        start = None
         end = today_event.start.date() - timedelta(days=1)
-        evens = get_events_in_time_frame(start, end, sender.id, session)
+        events = get_all_user_events(session, sender.id)
+        evens = filter_dates(events, start, end)
         assert list(evens) == []
 
     def test_get_events_start_time_frame_success(
@@ -51,8 +56,9 @@ class TestExport:
             sender, session
     ):
         start = next_month_event.start.date()
-        end = ''
-        evens = get_events_in_time_frame(start, end, sender.id, session)
+        end = None
+        events = get_all_user_events(session, sender.id)
+        evens = filter_dates(events, start, end)
         assert list(evens) == [next_month_event]
 
     def test_get_events_start_time_frame_failure(
@@ -61,8 +67,9 @@ class TestExport:
             sender, session
     ):
         start = next_month_event.start.date() + timedelta(days=1)
-        end = ''
-        evens = get_events_in_time_frame(start, end, sender.id, session)
+        end = None
+        events = get_all_user_events(session, sender.id)
+        evens = filter_dates(events, start, end)
         assert list(evens) == []
 
     def test_get_events_start_and_end_time_frame_success(
@@ -72,7 +79,8 @@ class TestExport:
     ):
         start = today_event.start.date()
         end = next_month_event.start.date()
-        evens = get_events_in_time_frame(start, end, sender.id, session)
+        events = get_all_user_events(session, sender.id)
+        evens = filter_dates(events, start, end)
         assert set(evens) == {today_event, next_month_event}
 
     def test_get_events_start_and_end_time_frame_failure(
@@ -82,7 +90,8 @@ class TestExport:
     ):
         start = today_event.start.date() + timedelta(days=1)
         end = next_month_event.start.date() - timedelta(days=1)
-        evens = get_events_in_time_frame(start, end, sender.id, session)
+        events = get_all_user_events(session, sender.id)
+        evens = filter_dates(events, start, end)
         assert list(evens) == []
 
     def test_create_ical_calendar(self):
